Refactor listen.py into smaller modules for better maintainability
